backend/requirements.txt
# ============================================================================
# Core Web Framework
# ============================================================================
fastapi==0.115.6
uvicorn[standard]==0.30.6
starlette>=0.27.0

# ============================================================================
# Database & ORM
# ============================================================================
sqlmodel==0.0.22
psycopg[binary]==3.2.3

# ============================================================================
# Pydantic & Validation
# ============================================================================
# OpenAI Agents SDK requires pydantic>=2.12.3
pydantic>=2.12.0,<3.0.0
pydantic-settings>=2.7.0

# ============================================================================
# Environment & Configuration
# ============================================================================
python-dotenv==1.0.0

# ============================================================================
# HTTP Client
# ============================================================================
httpx>=0.25.0

# ============================================================================
# Authentication & Security
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9
cryptography==42.0.8

# ============================================================================
# Rate Limiting
# ============================================================================
slowapi==0.1.8

# ============================================================================
# AI & Agents
# ============================================================================
# OpenAI Agents SDK (requires pydantic>=2.12.3)
openai-agents>=0.7.0
# OpenAI Python client for Gemini API compatibility


# ============================================================================
# MCP (Model Context Protocol)
# ============================================================================
# Official MCP Python SDK
mcp>=1.1.0

# ============================================================================
# Streaming (Server-Sent Events)
# ============================================================================
sse-starlette>=1.8.0
